Ensure that SWIG_VERSION is always defined
It should be defined both at SWIG-time and in the generated C/C++
wrapper code.

Also stop defining SWIGVERSION in the wrapper.  This only happened as a
side-effect of how SWIG_VERSION was defined but was never documented and
is redundant.

WIP: A number of languages don't currently define SWIG_VERSION in the
generated wrapper code.